+++ ccc-20200701/src/Models/AppendCasesResponseBody.php
@@ -5,6 +5,7 @@
 namespace AlibabaCloud\SDK\CCC\V20200701\Models;
 
 use AlibabaCloud\Dara\Model;
+use AlibabaCloud\SDK\CCC\V20200701\Models\AppendCasesResponseBody\data;
 
 class AppendCasesResponseBody extends Model
 {
@@ -12,6 +13,11 @@ class AppendCasesResponseBody extends Model
      * @var string
      */
     public $code;
+
+    /**
+     * @var data[]
+     */
+    public $data;
 
     /**
      * @var string
@@ -29,6 +35,7 @@ class AppendCasesResponseBody extends Model
     public $requestId;
     protected $_name = [
         'code' => 'Code',
+        'data' => 'Data',
         'httpStatusCode' => 'HttpStatusCode',
         'message' => 'Message',
         'requestId' => 'RequestId',
@@ -36,6 +43,9 @@ class AppendCasesResponseBody extends Model
 
     public function validate()
     {
+        if (\is_array($this->data)) {
+            Model::validateArray($this->data);
+        }
         parent::validate();
     }
 
@@ -44,6 +54,17 @@ class AppendCasesResponseBody extends Model
         $res = [];
         if (null !== $this->code) {
             $res['Code'] = $this->code;
+        }
+
+        if (null !== $this->data) {
+            if (\is_array($this->data)) {
+                $res['Data'] = [];
+                $n1 = 0;
+                foreach ($this->data as $item1) {
+                    $res['Data'][$n1] = null !== $item1 ? $item1->toArray($noStream) : $item1;
+                    ++$n1;
+                }
+            }
         }
 
         if (null !== $this->httpStatusCode) {
@@ -71,6 +92,17 @@ class AppendCasesResponseBody extends Model
         $model = new self();
         if (isset($map['Code'])) {
             $model->code = $map['Code'];
+        }
+
+        if (isset($map['Data'])) {
+            if (!empty($map['Data'])) {
+                $model->data = [];
+                $n1 = 0;
+                foreach ($map['Data'] as $item1) {
+                    $model->data[$n1] = data::fromMap($item1);
+                    ++$n1;
+                }
+            }
         }
 
         if (isset($map['HttpStatusCode'])) {